Move screens related code into its own folder
