+++ crates/lumina_client/src/ui.rs
@@ -2,6 +2,8 @@ use bevy::prelude::*;
 use bevy_radiance_cascades::prelude::*;
 use bevy_vello::render::VelloCanvasMaterial;
 
+use crate::screens::Screen;
+
 use super::Connection;
 
 pub(super) mod game_mode;
@@ -15,7 +17,7 @@ pub(super) struct UiPlugin;
 
 impl Plugin for UiPlugin {
     fn build(&self, app: &mut App) {
-        app.init_state::<Screen>().add_plugins((
+        app.add_plugins((
             splash::SplashUiPlugin,
             main_menu::MainMenuUiPlugin,
             game_mode::GameModeUiPlugin,
@@ -34,6 +36,7 @@ fn return_to_main_menu(mut next_screen_state: ResMut<NextState<Screen>>) {
     next_screen_state.set(Screen::MainMenu);
 }
 
+// Set all vello canvas to be ignored by radiancs cascades global illumination.
 fn set_no_radiance(
     mut commands: Commands,
     q_scenes: Query<Entity, (With<Handle<VelloCanvasMaterial>>, Without<NoRadiance>)>,
@@ -41,22 +44,4 @@ fn set_no_radiance(
     for entity in q_scenes.iter() {
         commands.entity(entity).insert(NoRadiance);
     }
-}
-
-#[derive(States, Debug, Hash, PartialEq, Eq, Clone, Default)]
-pub enum Screen {
-    // #[default]
-    Splash,
-    #[default]
-    MainMenu,
-    LocalLobby,
-    Sandbox,
-    Matchmaking,
-    MultiplayerLobby,
-    InGame,
-    GameOver,
-    // Leaderboard,
-    // Tutorial,
-    // Credits,
-    // Loading,
 }
